wv: sync-up plugin (fix-up)
3137805 Merge remote-tracking branch 'ohd/oc-widevine-release' into HEAD

This update picks up commit 43b8522 that fixes a problem that mediashell
would lock up if a content is casted twice in Google Play Movies from
another device.

The previous update (55ac41c) was mistakenly using the previous widevine
baseline, so it did not actually pick up the change quoted above.  This
is corrected now.

Based on 2018-05-06_23-00-02.

Change-Id: I14edb2d283a6a448efe001c3784024aed851d90e